generate openapi doc via artisan command
